refactor: split public contract modules
