+++ src/ports/security.ts
@@ -1,0 +1,4 @@
+export interface PasswordHasher {
+  hash(password: string): Promise<string>
+  verify(password: string, passwordHash: string): Promise<boolean>
+}
